backend/pyproject.toml
[project]
name = "epidemiologia-api"
version = "1.0.0"
description = "API del Sistema de Epidemiología"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0.0,<2.1.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "pandas>=2.0.0",
    "polars>=1.0.0",
    "pyarrow>=17.0.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "python-calamine>=0.1.0",
    "fastexcel>=0.11.0",
    "python-magic>=0.4.27",
    "asyncpg>=0.30.0",
    "werkzeug>=3.0.0",
    "click>=8.1.0",
    "greenlet>=3.2.3",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "svglib>=1.5.0",
    "xhtml2pdf>=0.2.11",
    "python-docx>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4,<2.0",
    "requests>=2.31.0",
    "geoalchemy2>=0.14.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.3.0",
    "httpx>=0.27.0",
    "testcontainers>=4.0.0",
]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.7",
    "ty>=0.0.1a32",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Configuración de ty (type checker)
[tool.ty.environment]
python-version = "3.11"

[tool.ty.src]
include = ["app"]
exclude = ["alembic", "tests", "data"]

# Reglas: ignorar falsos positivos comunes de SQLAlchemy/SQLModel/Starlette
[tool.ty.rules]
# SQLAlchemy/SQLModel genera muchos falsos positivos por metaprogramming
possibly-unresolved-reference = "warn"
unresolved-attribute = "warn"
# deprecated: session.execute() es la forma correcta para queries multi-columna
# aunque SQLModel lo marca deprecated a favor de exec() que no soporta tuplas
deprecated = "ignore"

# Overrides para archivos específicos con falsos positivos de Starlette middleware
[[tool.ty.overrides]]
include = ["app/main.py", "app/core/middleware.py"]

[tool.ty.overrides.rules]
# Starlette usa ParamSpec que ty no maneja bien con add_middleware
invalid-argument-type = "ignore"

